refactor(DashboardLayout): Move data request functions and combine states
