+++ src/routes/Players.jsx
@@ -1,4 +1,4 @@
-import { useEffect, useState } from 'react';
+import { useEffect, useMemo, useState } from 'react';
 import { useOutletContext } from 'react-router-dom';
 import DashboardSummary from '../components/dashboard/DashboardSummary';
 import styled from 'styled-components';
@@ -6,7 +6,6 @@ import Searchbar from '../components/common/Searchbar';
 import { BanIcon, KickIcon } from '../assets/24x-icons';
 import {
   currentProfileState,
-  playerDetailState,
   playersState
 } from '../contexts/states';
 import { useRecoilState, useRecoilValue } from 'recoil';
@@ -130,8 +129,8 @@ const PlayerButton = styled.button`
 `;
 
 const PlayerInfoContainer = ({ uuid, name }) => {
-  const [playerDetails, setPlayerDetails] = useRecoilState(playerDetailState);
-  const player = playerDetails[uuid];
+  const [players, setPlayers] = useRecoilState(playersState);
+  const player = players[uuid];
   const profile = useRecoilValue(currentProfileState);
 
   return player ? (
@@ -164,7 +163,7 @@ const PlayerInfoContainer = ({ uuid, name }) => {
               { reason: '' }
             );
 
-            setPlayerDetails((details) => ({
+            setPlayers((details) => ({
               ...details,
               [uuid]: undefined
             }));
@@ -185,7 +184,7 @@ const PlayerInfoContainer = ({ uuid, name }) => {
               { reason: '' }
             );
 
-            setPlayerDetails((details) => ({
+            setPlayers((details) => ({
               ...details,
               [uuid]: undefined
             }));
@@ -210,7 +209,8 @@ const Players = () => {
     label: '이름'
   });
   const [searchValue, setSearchValue] = useState('');
-  const players = useRecoilValue(playersState);
+  const playerDetails = useRecoilValue(playersState);
+  const players = useMemo(() => Object.values(playerDetails), [playerDetails])
 
   useEffect(() => {
     // 이 컴포넌트에서 DashboardLayout으로 정보 새로 고침 함수를 넘겨야 합니다
